requirements.txt
fastapi
uvicorn
langchain
langgraph
langchain-openai
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
python-dotenv
httpx
